+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Advantages:
 - [macOS]: Added the [plist configuration guide](README-MACOS-PLIST.md).
 - [frontend]: Added download diagnostic and download history to Download menu.
 - [frontend]: Added icon to open the cpu and memory usage in System Information panel.
+- [thermostat]: Added thermostatRunningState. Thanks Ludovic BOUÉ (https://github.com/Luligu/matterbridge/pull/410).
 
 ### Changed